LayoutTests/media/media-source/media-source-cea608simple-expected.txt

media-source-cea608simple

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)

Append media segments.
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(1)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(2)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(3)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(4)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(5)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(6)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(7)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(8)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(9)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(10)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(11)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(12)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(13)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(14)))
EVENT(update)

Verify track properties.
EXPECTED (video.videoTracks.length == '1') OK
EXPECTED (video.audioTracks.length == '0') OK
EXPECTED (video.textTracks.length == '1') OK
RUN(track = video.textTracks[0])
EXPECTED (track.id == '1') OK

Start playback and verify cues
RUN(track.mode = "showing")
EVENT(change)
RUN(cues = track.cues)
EXPECTED (cues.length == '30') OK
RUN(video.play())
RUN(video.pause())
EXPECTED (cues[0].text == 'The screen should show number 1.') OK
EXPECTED (track.activeCues.length == '1') OK
END OF TEST
